% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi_prob.R
\name{EquiProb}
\alias{EquiProb}
\title{Practical Equivalence Probability}
\usage{
EquiProb(cens_prop, n, rate1, rate2, shape1, shape2, margin = 0, reps = 10)
}
\arguments{
\item{cens_prop}{Expected censoring proportion.}

\item{n}{Sample size.}

\item{rate1}{Rate parameter for arm 1.}

\item{rate2}{Rate parameter for arm 2.}

\item{shape1}{Shape parameter for arm 1.}

\item{shape2}{Shape parameter for arm 2.}

\item{margin}{Equivlence margin.}

\item{reps}{Simulation replicates.}
}
\value{
Numeric equivalence probability.
}
\description{
Equivalence probability, as pr(median2 - median1 >= margian) 
+ 0.5 * pr(abs(median2 - median1) < margin).
}
