% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_ordering_prob.R
\name{OrderProb}
\alias{OrderProb}
\title{Calculate Ordering Probability}
\usage{
OrderProb(n, rate1, rate2, med1 = NULL, med2 = NULL, censor_prop = 0)
}
\arguments{
\item{n}{Sample size, assuming 1:1 randomization.}

\item{rate1}{True rate parameter for arm 1.}

\item{rate2}{True rate parameter for arm 2.}

\item{med1}{True median of arm 1, optionally supplied instead of rate1.}

\item{med2}{True median of arm 2, optionally supplied instead of rate2.}

\item{censor_prop}{Expected censoring proportion.}
}
\value{
Numeric probability.
}
\description{
Calculates the asymptotic probability that the estimated rate parameter for arm 1
is less than the estimated rate parameter for arm 2.
}
\examples{
prob <- OrderProb(n = 10, med1 = 8, med2 = 6)
prob <- OrderProb(n = 50, med1 = 8, med2 = 6)
}
