% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_prob.R
\name{WeiAvgInfo}
\alias{WeiAvgInfo}
\title{Weibull Average Information}
\usage{
WeiAvgInfo(cens_prop, n, shape, rate, reps = 10)
}
\arguments{
\item{cens_prop}{Censoring proportion.}

\item{n}{Sample size.}

\item{shape}{Shape parameter `alpha`.}

\item{rate}{Rate parameter `lambda`.}

\item{reps}{Replicates to average.}
}
\value{
Numeric information matrix for shape and rate.
}
\description{
Estimate the expected information as the average value of the observed
information across `reps` realizations of the data.
}
