% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_specification.R
\name{WeibullSpec}
\alias{WeibullSpec}
\title{Weibull Specification}
\usage{
WeibullSpec(t1, p1, t2, p2)
}
\arguments{
\item{t1}{First time point.}

\item{p1}{Probability at the first time point.}

\item{t2}{Second time point.}

\item{p2}{Probability at the send time point.}
}
\value{
Numeric vector containing the shape and rate.
}
\description{
Calculate shape and rate of weibull distribution from two quantiles.
}
\examples{
theta <- WeibullSpec(t1 = 6, p1 = 0.8, t2 = 12, p2 = 0.5)
}
