% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi_prob.R
\name{WeiInfo}
\alias{WeiInfo}
\title{Weibull Information Matrix.}
\usage{
WeiInfo(data, shape, rate)
}
\arguments{
\item{data}{Data.frame.}

\item{shape}{Shape parameter, alpha.}

\item{rate}{Rate parameter, lambda.}
}
\value{
Numeric information matrix for shape and rate.
}
\description{
Information matrix for the Weibull shape and rate parameters.
}
