% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_prob.R
\name{WeiMed}
\alias{WeiMed}
\title{Weibull Median}
\usage{
WeiMed(shape, rate)
}
\arguments{
\item{shape}{Shape parameter, `alpha`.}

\item{rate}{Rate parameter, `lambda`.}
}
\value{
Numeric median.
}
\description{
Weibull Median
}
