% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi_prob.R
\name{WeiMedSE}
\alias{WeiMedSE}
\title{Weibull Median SE}
\usage{
WeiMedSE(info, n, shape, rate)
}
\arguments{
\item{info}{Information matrix.}

\item{n}{Sample size.}

\item{shape}{Shape parameter `alpha`.}

\item{rate}{Rate parameter `lambda`.}
}
\value{
Numeric standard error of the median.
}
\description{
Weibull Median SE
}
