% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_calc.R
\name{ProbCurve}
\alias{ProbCurve}
\title{Probability by Sample Size Curve.}
\usage{
ProbCurve(rate1, rate2, censor_prop = 0, min_n = 1, max_n = 100)
}
\arguments{
\item{rate1}{True rate parameter for arm 1.}

\item{rate2}{True rate parameter for arm 2.}

\item{censor_prop}{Expected censoring proportion.}

\item{min_n}{Minimum allowable sample size.}

\item{max_n}{Maximum allowable sample size.}
}
\value{
Data.frame.
}
\description{
Constructs a data.frame containing the probability that the rate 
parameter for arm 1 is <= the rate parameter for arm 2 for 
sample size between min_n and max_n.
}
