% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_calc.R
\name{SampleSize}
\alias{SampleSize}
\title{Calculate Sample Size}
\usage{
SampleSize(
  rate1,
  rate2,
  med1 = NULL,
  med2 = NULL,
  censor_prop = 0,
  min_n = 1,
  max_n = 100,
  target_prob = 0.9
)
}
\arguments{
\item{rate1}{True rate parameter for arm 1.}

\item{rate2}{True rate parameter for arm 2.}

\item{med1}{True median of arm 1, optionally supplied instead of rate1.}

\item{med2}{True median of arm 2, optionally supplied instead of rate2.}

\item{censor_prop}{Expected censoring proportion.}

\item{min_n}{Minimum allowable sample size.}

\item{max_n}{Maximum allowable sample size.}

\item{target_prob}{Probability of selecting the more-effective treatment.}
}
\value{
Integer sample size.
}
\description{
Calculates the sample size needed to provide a specified probability of
selecting the more-effective treatment.
}
\examples{
# Sample size calculation, specified in terms of rates.
n <- SampleSize(rate1 = 1.0, rate2 = 1.4)

# Sample size calculation, specified in terms of medians.
n <- SampleSize(med1 = 12, med2 = 8)
}
