% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_prob.R
\name{WeiMed}
\alias{WeiMed}
\title{Weibull Median}
\usage{
WeiMed(shape, rate)
}
\arguments{
\item{shape}{Shape parameter, `alpha`.}

\item{rate}{Rate parameter, `lambda`.}
}
\value{
Numeric median.
}
\description{
Calculate the median of a Weibull distribution from the shape and rate.
}
\examples{
# In the case of shape = 1 and rate = 1, the distribution
# is exponential, and the median is log(2).
med <- WeiMed(shape = 1, rate = 1)
}
